% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ref}
\alias{ref}
\alias{ref<-}
\title{Extract and replace the "ref" form \code{twLink} and \code{twImage} objects.}
\usage{
ref(object)
}
\description{
Generic function used to access the "ref" slot in \code{twLink} and \code{twImage} objects.
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{twLink}}, \code{\link{twImage}}
}
\keyword{image}
\keyword{link}
\keyword{ref}
\keyword{wiki}

