\docType{package}
\name{TiddlyWikiR-package}
\alias{TiddlyWikiR}
\alias{TiddlyWikiR-package}
\title{TiddlyWikiR: an R package for writing dynamic reports using a TiddlyWiki template.}
\description{
  TiddlyWikiR is an R package for writing dynamic reports
  using TiddlyWiki as a template. The library implements S4
  classes to organize and handle statistical results within
  R, and functions to insert those results into the wiki
  file.

  A quick overview of the functionality of the library can
  be found in the HTML vignette. You can access to it by
  typing \code{browseVignettes ("TiddlyWikiR")}

  Visit https://github.com/dmontaner/TiddlyWikiR for the
  most updated version of the package.

  More documentation on TiddlyWikiR may be found at
  http://www.dmontaner.com/

  More information about TiddlyWiki is available at
  http://tiddlywiki.comk
}
\author{
  David Montaner \email{dmontaner@cipf.es}
}

