\name{twBold}
\alias{twBold}
\alias{twCode}
\alias{twHigh}
\alias{twIt}
\alias{twUnder}
\title{Basic TiddlyWiki formatting functions}
\usage{
  twBold(x)
}
\arguments{
  \item{x}{text to be formatted}
}
\description{
  Some basic functions for basic formatting.
}
\details{
  See http://classic.tiddlywiki.com/#[[Basic Formatting]]
  for further details.
}
\section{Usage}{
  twBold (x) twIt (x) twUnder (x) twHigh (x) twCode (x)
}
\examples{
twBold ("here my bold")
twCode ("here my code")
}
\author{
  David Montaner \email{dmontaner@cipf.es}
}
\seealso{
  \code{\link{twImage}} and \code{\link{twLink}}
}
\keyword{formats}
\keyword{wiki}

