\docType{methods}
\name{align}
\alias{align}
\alias{align<-}
\title{Extract and replace the "alignment" form \code{twTable} and \code{twImage} objects.}
\description{
  Generic function used to access the "align" slot in
  \code{twTable} and \code{twImage} objects.
}
\author{
  David Montaner \email{dmontaner@cipf.es}
}
\seealso{
  \code{\link{twTable}}, \code{\link{twImage}}
}
\keyword{align}
\keyword{image}
\keyword{table}
\keyword{wiki}

