\docType{methods}
\name{label}
\alias{label}
\alias{label<-}
\title{Extract and replace the "label" form \code{twLink} and \code{twImage} objects.}
\description{
  Generic function used to access the "label" slot in
  \code{twLink} and \code{twImage} objects.
}
\author{
  David Montaner \email{dmontaner@cipf.es}
}
\seealso{
  \code{\link{twLink}}, \code{\link{twImage}}
}
\keyword{image}
\keyword{label}
\keyword{link}
\keyword{wiki}

